escape-room-forecast/pyproject.toml
[tool.poetry]
name = "escape-room-forecast"
version = "0.1.0"
description = "Forecasting for escape room bookings"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "forecast_cli", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
torch = {version = "2.3.1+cu121", source = "pytorch_cuda"}
autogluon-timeseries = {extras = ["gpu"], version = "^1.3"}
# For Mac compatibility (conversion)
coremltools = {version = "^7.1", optional = true}
polars = {version = "^1.0.0", optional = true}
pyyaml = "^6.0.1" # Explicitly add for YAML config loading
workalendar = "^17.0.0" # Added for holiday features
typer = {extras = ["all"], version = "^0.12.3"}
pandas = "^2.2.3"
openmeteo-requests = "^1.4.0"
retry-requests = "^2.0.0"
requests-cache = "^1.2.1"
msgpack = "^1.1.0"

[[tool.poetry.source]]
name = "pytorch_cuda"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
ruff = "^0.5.0"
black = "^24.4.2"
mypy = "^1.10.0"
pre-commit = "^3.7.1"

[tool.poetry.extras]
gpu = ["torch"]
conversion = ["coremltools", "polars"]

[tool.poetry.scripts]
forecast = "forecast_cli.cli:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Tool configurations
[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "UP", "C4", "B", "A", "RUF"]
ignore = ["E501"]

[tool.ruff.lint]
# Explicitly configure per-file ignores if necessary later
# per-file-ignores = { "**/__init__.py" = ["F401"] }

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
# plugins = [
#   "pydantic.mypy" # If using pydantic later
# ] 